Add hotjar function as plugin
